+++ Final/Learners/ViewModels/LearnerViewModel.swift
@@ -1,0 +1,72 @@
+//
+//  LearnerViewModel.swift
+//  Learners
+//
+//  Created by Luca Palmese on 16/10/23.
+//
+
+import Foundation
+
+class LearnerViewModel {
+    
+    var learners = [
+        Learner(name: "Alex", surname: "Doe", favouriteColor: .pink, description: "I love myself", imageName: "adriano"),
+        Learner(name: "Chris", surname: "Smith", favouriteColor: .blue, description: "I'm strong", imageName: "alessandro"),
+        Learner(name: "James", surname: "Johnson", favouriteColor: .red, description: "I'm strong", imageName: "alessandro"),
+        Learner(name: "Michael", surname: "Brown", favouriteColor: .yellow,  description: "I'm strong", imageName: "alessandro"),
+        Learner(name: "Jessica", surname: "White", favouriteColor: .orange,  description: "I'm invincible", imageName: "alessia"),
+        Learner(name: "Daniel", surname: "Miller", favouriteColor: .green),
+        Learner(name: "Emily", surname: "Davis", favouriteColor: .pink),
+        Learner(name: "Sophia", surname: "Garcia", favouriteColor: .black),
+        Learner(name: "Matthew", surname: "Martinez", favouriteColor: .blue),
+        Learner(name: "Joshua", surname: "Hernandez", favouriteColor: .red),
+        Learner(name: "Andrew", surname: "Lopez", favouriteColor: .orange),
+        Learner(name: "Brandon", surname: "Gonzalez", favouriteColor: .yellow),
+        Learner(name: "Olivia", surname: "Wilson", favouriteColor: .green),
+        Learner(name: "Isabella", surname: "Anderson", favouriteColor: .black),
+        Learner(name: "Aiden", surname: "Thomas", favouriteColor: .pink),
+        Learner(name: "Caleb", surname: "Taylor", favouriteColor: .blue),
+        Learner(name: "Ethan", surname: "Moore", favouriteColor: .red),
+        Learner(name: "Mia", surname: "Jackson", favouriteColor: .orange),
+        Learner(name: "Abigail", surname: "Martin", favouriteColor: .orange),
+        Learner(name: "Lucas", surname: "Lee", favouriteColor: .black),
+        Learner(name: "Nathan", surname: "Perez", favouriteColor: .pink),
+        Learner(name: "Liam", surname: "Thompson", favouriteColor: .brown),
+        Learner(name: "Noah", surname: "White", favouriteColor: .blue),
+        Learner(name: "Logan", surname: "Harris", favouriteColor: .red),
+        Learner(name: "Avery", surname: "Clark", favouriteColor: .orange),
+        Learner(name: "Samantha", surname: "Lewis", favouriteColor: .yellow),
+        Learner(name: "Victoria", surname: "Robinson", favouriteColor: .black),
+        Learner(name: "Mason", surname: "Walker", favouriteColor: .green),
+        Learner(name: "Ella", surname: "Young", favouriteColor: .black),
+        Learner(name: "Sebastian", surname: "Allen", favouriteColor: .pink),
+        Learner(name: "Ryan", surname: "King", favouriteColor: .red),
+        Learner(name: "Charlotte", surname: "Scott", favouriteColor: .blue),
+        Learner(name: "Jack", surname: "Green", favouriteColor: .yellow),
+        Learner(name: "Zachary", surname: "Baker", favouriteColor: .green),
+        Learner(name: "Jason", surname: "Adams", favouriteColor: .red),
+        Learner(name: "Connor", surname: "Nelson", favouriteColor: .black),
+        Learner(name: "Grace", surname: "Carter", favouriteColor: .green),
+        Learner(name: "Leah", surname: "Mitchell", favouriteColor: .pink),
+        Learner(name: "Hailey", surname: "Perez", favouriteColor: .blue),
+        Learner(name: "Lily", surname: "Roberts", favouriteColor: .brown),
+        Learner(name: "Sarah", surname: "Hall", favouriteColor: .yellow),
+        Learner(name: "Zoe", surname: "Clark", favouriteColor: .black),
+        Learner(name: "Madison", surname: "Rodriguez", favouriteColor: .green),
+        Learner(name: "Natalie", surname: "Lewis", favouriteColor: .yellow),
+        Learner(name: "Sophie", surname: "Walker", favouriteColor: .black),
+        Learner(name: "Max", surname: "Young", favouriteColor: .orange),
+        Learner(name: "Maya", surname: "Allen", favouriteColor: .blue),
+        Learner(name: "Emma", surname: "King", favouriteColor: .red),
+        Learner(name: "Chloe", surname: "Wright", favouriteColor: .green),
+        Learner(name: "Ella", surname: "Scott", favouriteColor: .blue),
+        Learner(name: "Isla", surname: "Green", favouriteColor: .pink),
+        Learner(name: "Jaxon", surname: "Adams", favouriteColor: .pink),
+        Learner(name: "Harper", surname: "Nelson", favouriteColor: .pink),
+        Learner(name: "Daniel", surname: "Carter", favouriteColor: .red),
+        Learner(name: "Samuel", surname: "Moore", favouriteColor: .blue),
+        Learner(name: "Henry", surname: "Brown", favouriteColor: .yellow),
+        Learner(name: "Matthew", surname: "Williams", favouriteColor: .black),
+        Learner(name: "Elijah", surname: "Miller", favouriteColor: .orange)
+    ]
+}