+++ docs/CHANGELOG.md
@@ -1,7 +1,8 @@
 ## CHANGELOG
 
-## Version 1.0.0 (aaaa-mm-dd)
+## Version 1.0.0 (2024-09-18)
 
 First release:
 
-* Activity-Template - <Comment>
+* Lab Navigation in SwiftUI - Starter
+* Lab Navigation in SwiftUI - Final